Disease list is retrived successfuly

+++ Flutter-Harpia-Health-Analysis/lib/model/user/Admin.dart
@@ -5,7 +5,7 @@ import 'User.dart';
 
 class Admin extends User {
   Admin(
-      {required Long id,
+      {required int id,
       required int roleId,
       required String name,
       required String lastname,
@@ -28,7 +28,7 @@ class Admin extends User {
 
   factory Admin.fromJson(Map<String, dynamic> json) {
     return Admin(
-        id: json["id"] as Long,
+        id: json["id"] as int,
         roleId: json["roleId"] as int,
         name: json["name"] as String,
         lastname: json['lastname'] as String,
